* tag 'v3.4.0': (24 commits)
  Publish
  Prepare 3.4.0 release
  Downgrade style-loader to v0.23.1 due to CSS modules hot reload… (#8378)
  Correct webpack name casing (#8475)
  update open to v7.0.2 (#8459)
  fix(typescriptFormatter): use chalk@2 constructor (#8450)
  fix(react-scripts): do not redirect served path if request may proxy (#8442)
  Downgrade chalk for ie 11 support (#8439)
  Fixes unchecked access to 'deploy' script on build (#8292)
  Fix navbar line break in header (#8437)
  Change arrow functions to function declarations (#8412)
  feat(react-scripts): allow PUBLIC_URL in develoment mode (#7259)
  Handle service worker error in Firefox (#8272)
  Add option to provide custom ssl certificates during development (#5845)
  Enable custom sockjs pathname for hot reloading server. (#7750)
  Support JetBrains Rider IDE as an editor (#7948)
  Run git init before template dependencies are installed (#8282)
  Update public folder usage docs to clarify globals (#8299)
  Fix robots.txt for TS (#8403)
  setupTestFrameworkScriptFile is deprecated (#8390)
  ...

+++ packages/react-scripts/scripts/init.js
@@ -64,20 +64,22 @@ function tryGitCommit(appPath) {
         });
         return true;
     } catch (e) {
-        // We couldn't commit in already initialized git repo,
-        // maybe the commit author config is not set.
-        // In the future, we might supply our own committer
-        // like Ember CLI does, but for now, let's just
-        // remove the Git files to avoid a half-done state.
-        console.warn('Git commit not created', e);
-        console.warn('Removing .git directory...');
-        try {
-            // unlinkSync() doesn't work on directories.
-            fs.removeSync(path.join(appPath, '.git'));
-        } catch (removeErr) {
-            // Ignore.
+        if (didInit) {
+            // If we successfully initialized but couldn't commit,
+            // maybe the commit author config is not set.
+            // In the future, we might supply our own committer
+            // like Ember CLI does, but for now, let's just
+            // remove the Git files to avoid a half-done state.
+            console.warn('Git commit not created', e);
+            console.warn('Removing .git directory...');
+            try {
+                // unlinkSync() doesn't work on directories.
+                fs.removeSync(path.join(appPath, '.git'));
+            } catch (removeErr) {
+                // Ignore.
+            }
+            return false;
         }
-        return false;
     }
 }
 
@@ -396,18 +398,6 @@ function installRioDevDependencies(useYarn, verbose, templateJson) {
     }
 
     return proc.status;
-}
-
-function tryFinalGitAdd(appPath) {
-    try {
-        execSync('git add -A', { stdio: 'ignore' });
-        execSync('git commit -m "Initial commit from Create React App"', {
-            stdio: 'ignore',
-        });
-        return true;
-    } catch (e) {
-        return false;
-    }
 }
 
 function getRioBrowserList() {
